+++ source/homebuttonLib/nw4hbm/math/types.h
@@ -7,7 +7,7 @@
 
 #include <revolution/types.h>
 
-// #include <revolution/mtx.h>
+#include <revolution/mtx.h>
 
 #include <context_rvl.h>
